refactor list
