+++ packages/datatrak-web/src/features/MobileSelectList/index.ts
@@ -1,0 +1,6 @@
+/*
+ * Tupaia
+ *  Copyright (c) 2017 - 2024 Beyond Essential Systems Pty Ltd
+ */
+
+export { MobileSelectList } from './MobileSelectList';
